requirements.txt
asgiref==3.2.10
astroid==2.4.2
beautifulsoup4==4.9.1
certifi==2020.6.20
cffi==1.14.1
chardet==3.0.4
colorama==0.4.3
cryptography==3.0
cssutils==1.0.2
defusedxml==0.6.0
Django==3.0.8
django-allauth==0.44.0
django-analytical==3.0.0
django-ckeditor==6.0.0
django-crispy-forms==1.9.2
django-groupadmin-users==0.3.3
django-inlinecss==0.3.0
django-jquery==3.1.0
django-js-asset==1.2.2
future==0.18.2
idna==2.10
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
oauthlib==3.1.0
Pillow==7.2.0
pycparser==2.20
PyJWT==1.7.1
pylint==2.5.3
pynliner==0.8.0
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
toml==0.10.1
urllib3==1.25.10
wrapt==1.12.1
